[YW2-79] naming 변경 : volunteer -> user
