+++ src/app/api/src/main/kotlin/yapp/be/apiapplication/auth/controller/model/LoginVolunteerRequest.kt
@@ -1,0 +1,16 @@
+package yapp.be.apiapplication.auth.controller.model
+
+import yapp.be.apiapplication.auth.service.model.LoginVolunteerRequestDto
+import yapp.be.model.Email
+
+data class LoginVolunteerRequest(
+    val email: String,
+    val authCode: String,
+) {
+    fun toDto(): LoginVolunteerRequestDto {
+        return LoginVolunteerRequestDto(
+            email = Email(email),
+            authCode = authCode
+        )
+    }
+}
