--- src/app/api/src/main/kotlin/yapp/be/apiapplication/auth/controller/model/UserSignUpRequest.kt
@@ -1,18 +1,0 @@
-package yapp.be.apiapplication.auth.controller.model
-
-import yapp.be.apiapplication.auth.service.model.SignUpUserRequestDto
-import yapp.be.model.Email
-
-data class UserSignUpRequest(
-    val nickname: String,
-    val email: String,
-    val phone: String,
-) {
-    fun toDto(): SignUpUserRequestDto {
-        return SignUpUserRequestDto(
-            email = Email(email),
-            nickname = nickname,
-            phone = phone,
-        )
-    }
-}
