--- src/domain/user/src/main/kotlin/yapp.be.domain/port/inbound/CheckUserUseCase.kt
@@ -1,6 +1,0 @@
-package yapp.be.domain.port.inbound
-
-interface CheckUserUseCase {
-    fun isExistByEmail(email: String): Boolean
-    fun isExistByNickname(nickname: String): Boolean
-}
