--- src/domain/user/src/main/kotlin/yapp.be.domain/port/inbound/GetUserUseCase.kt
@@ -1,7 +1,0 @@
-package yapp.be.domain.port.inbound
-
-import yapp.be.domain.model.User
-
-interface GetUserUseCase {
-    fun getByEmail(email: String): User
-}
