--- src/domain/user/src/main/kotlin/yapp.be.domain/port/inbound/ModifyUserUseCase.kt
@@ -1,8 +1,0 @@
-package yapp.be.domain.port.inbound
-
-import yapp.be.domain.model.User
-import yapp.be.domain.port.inbound.model.ModifyUserTokenCommand
-
-interface ModifyUserUseCase {
-    fun updateUserToken(command: ModifyUserTokenCommand): User
-}
