+++ src/domain/volunteer/src/main/kotlin/yapp.be.domain/port/inbound/CheckVolunteerUseCase.kt
@@ -1,0 +1,6 @@
+package yapp.be.domain.port.inbound
+
+interface CheckVolunteerUseCase {
+    fun isExistByEmail(email: String): Boolean
+    fun isExistByNickname(nickname: String): Boolean
+}
