+++ src/domain/volunteer/src/main/kotlin/yapp.be.domain/port/inbound/CreateVolunteerUseCase.kt
@@ -1,0 +1,8 @@
+package yapp.be.domain.port.inbound
+
+import yapp.be.domain.model.Volunteer
+import yapp.be.domain.port.inbound.model.CreateUserCommand
+
+interface CreateVolunteerUseCase {
+    fun create(command: CreateUserCommand): Volunteer
+}
