+++ src/domain/volunteer/src/main/kotlin/yapp.be.domain/port/inbound/GetVolunteerUseCase.kt
@@ -1,0 +1,7 @@
+package yapp.be.domain.port.inbound
+
+import yapp.be.domain.model.Volunteer
+
+interface GetVolunteerUseCase {
+    fun getByEmail(email: String): Volunteer
+}
