+++ src/domain/volunteerEvent/src/main/kotlin/yapp.be.domain/model/VolunteerEvent.kt
@@ -14,6 +14,6 @@ data class VolunteerEvent(
     val viewCnt: Int,
     val volunteerEventStatus: VolunteerEventStatus,
     val participantNum: Int,
-    val userId: Long,
+    val volunteerId: Long,
     val shelterId: Long,
 )
