+++ src/domain/volunteerEvent/src/main/kotlin/yapp.be.domain/model/VolunteerEventJoinQueue.kt
@@ -2,6 +2,6 @@ package yapp.be.domain.model
 
 data class VolunteerEventJoinQueue(
     val id: Long,
-    val userId: Long,
+    val volunteerId: Long,
     val volunteerEventId: Long,
 )
