+++ src/domain/volunteerEvent/src/main/kotlin/yapp.be.domain/model/VolunteerEventWaitingQueue.kt
@@ -2,6 +2,6 @@ package yapp.be.domain.model
 
 data class VolunteerEventWaitingQueue(
     val id: Long,
-    val userId: Long,
+    val volunteerId: Long,
     val volunteerEventId: Long,
 )
