--- src/infrastructure/storage/src/main/kotlin/yapp/be/storage/jpa/user/repository/UserJpaRepository.kt
@@ -1,11 +1,0 @@
-package yapp.be.storage.jpa.user.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-import yapp.be.storage.jpa.user.model.UserEntity
-
-@Repository
-interface UserJpaRepository : JpaRepository<UserEntity, Long> {
-    fun findByEmail(email: String): UserEntity?
-    fun findByNickname(nickname: String): UserEntity?
-}
